New functions for a different plotting way
Two new functions added. The first one takes a group of CSV files that contain the same type of data with the same column structure and puts the data from each CSV file in a dictionary as a pandas dataframe. The user can give keys as a tag for each dataframe and column titles for each dataframe as function arguments. The second function is the second version of the plotting, function plot_sepctrum. The new version, plot_spectrumv2, takes a dictionary consist of dataframes with tags, and can plot any data list from those dataframes.