Initial commit
This module is written to stabilize the spectrum from the measurement of a RhoC detector, on which I worked for my industrial internship. It creates a JSON file after a measurement with raw counts. The output is defined as raw because the data corresponds to the channels of the detector, and needs to be stabilized.