Initial commit
This module contains some functions required for footprint calculations, which are used to make necessary corrections in cases where the gamma-ray source is not a point source but is distributed throughout a volume. The functions in this module were meant to be used for a salt mine project. Therefore; some functions and constants may be needed to be changed for a given geometry.